+++ src/MedAppServer/src/test/kotlin/org/kert0n/medappserver/controller/UsingsControllerTest.kt
@@ -58,7 +58,7 @@ class UsingsControllerTest {
         val user = User(id = userId, hashedKey = "key")
         val medKit = MedKit(id = UUID.randomUUID())
         val drug = Drug(
-            id = drugId, name = "Drug", quantity = 100.0,
+            id = drugId, name = "Drug", quantity = 100.0.toBigDecimal(),
             quantityUnit = "mg", formType = null, category = null,
             manufacturer = null, country = null, description = null,
             medKit = medKit
@@ -67,7 +67,7 @@ class UsingsControllerTest {
             usingKey = UsingKey(userId, drugId),
             user = user,
             drug = drug,
-            plannedAmount = 30.0,
+            plannedAmount = 30.0.toBigDecimal(),
             createdAt = now,
             lastModified = now
         )
@@ -76,7 +76,7 @@ class UsingsControllerTest {
     private fun createTestUsingDTO(): UsingDTO = UsingDTO(
         userId = userId,
         drugId = drugId,
-        plannedAmount = 30.0,
+        plannedAmount = 30.0.toBigDecimal(),
         createdAt = now,
         lastModified = now
     )
@@ -95,7 +95,7 @@ class UsingsControllerTest {
             .andExpect(status().isOk)
             .andExpect(jsonPath("$").isArray)
             .andExpect(jsonPath("$[0].drugId").value(drugId.toString()))
-            .andExpect(jsonPath("$[0].plannedAmount").value(30.0))
+            .andExpect(jsonPath("$[0].plannedAmount").value(30.0.toBigDecimal()))
     }
 
     @Test
@@ -117,7 +117,7 @@ class UsingsControllerTest {
         )
             .andExpect(status().isOk)
             .andExpect(jsonPath("$.drugId").value(drugId.toString()))
-            .andExpect(jsonPath("$.plannedAmount").value(30.0))
+            .andExpect(jsonPath("$.plannedAmount").value(30.0.toBigDecimal()))
     }
 
     @Test
@@ -139,7 +139,7 @@ class UsingsControllerTest {
         whenever(usingService.createTreatmentPlan(eq(userId), any())).thenReturn(using)
         whenever(usingService.toUsingDTO(using)).thenReturn(dto)
 
-        val createDTO = UsingCreateDTO(drugId = drugId, plannedAmount = 30.0)
+        val createDTO = UsingCreateDTO(drugId = drugId, plannedAmount = 30.0.toBigDecimal())
 
         mockMvc.perform(
             post("/using")
@@ -148,7 +148,7 @@ class UsingsControllerTest {
                 .content(objectMapper.writeValueAsString(createDTO))
         )
             .andExpect(status().isCreated)
-            .andExpect(jsonPath("$.plannedAmount").value(30.0))
+            .andExpect(jsonPath("$.plannedAmount").value(30.0.toBigDecimal()))
     }
 
     @Test
@@ -156,7 +156,7 @@ class UsingsControllerTest {
         whenever(usingService.createTreatmentPlan(eq(userId), any()))
             .thenThrow(ResponseStatusException(HttpStatus.CONFLICT, "Already exists"))
 
-        val createDTO = UsingCreateDTO(drugId = drugId, plannedAmount = 30.0)
+        val createDTO = UsingCreateDTO(drugId = drugId, plannedAmount = 30.0.toBigDecimal())
 
         mockMvc.perform(
             post("/using")
@@ -169,12 +169,12 @@ class UsingsControllerTest {
 
     @Test
     fun `PUT update using - updates and returns using`() {
-        val using = createTestUsing().apply { plannedAmount = 50.0 }
-        val dto = createTestUsingDTO().copy(plannedAmount = 50.0)
+        val using = createTestUsing().apply { plannedAmount = 50.0.toBigDecimal() }
+        val dto = createTestUsingDTO().copy(plannedAmount = 50.0.toBigDecimal())
         whenever(usingService.updateTreatmentPlan(eq(userId), eq(drugId), any())).thenReturn(using)
         whenever(usingService.toUsingDTO(using)).thenReturn(dto)
 
-        val updateDTO = UsingUpdateDTO(plannedAmount = 50.0)
+        val updateDTO = UsingUpdateDTO(plannedAmount = 50.0.toBigDecimal())
 
         mockMvc.perform(
             put("/using/drug/$drugId")
@@ -183,17 +183,17 @@ class UsingsControllerTest {
                 .content(objectMapper.writeValueAsString(updateDTO))
         )
             .andExpect(status().isOk)
-            .andExpect(jsonPath("$.plannedAmount").value(50.0))
+            .andExpect(jsonPath("$.plannedAmount").value(50.0.toBigDecimal()))
     }
 
     @Test
     fun `POST record intake - records and returns using`() {
-        val using = createTestUsing().apply { plannedAmount = 20.0 }
-        val dto = createTestUsingDTO().copy(plannedAmount = 20.0)
-        whenever(usingService.recordIntake(eq(userId), eq(drugId), eq(10.0))).thenReturn(using)
+        val using = createTestUsing().apply { plannedAmount = 20.0.toBigDecimal() }
+        val dto = createTestUsingDTO().copy(plannedAmount = 20.0.toBigDecimal())
+        whenever(usingService.recordIntake(eq(userId), eq(drugId), eq(10.0.toBigDecimal()))).thenReturn(using)
         whenever(usingService.toUsingDTO(using)).thenReturn(dto)
 
-        val intakeRequest = IntakeRequest(quantityConsumed = 10.0)
+        val intakeRequest = IntakeRequest(quantityConsumed = 10.0.toBigDecimal())
 
         mockMvc.perform(
             post("/using/drug/$drugId/intake")
@@ -202,7 +202,7 @@ class UsingsControllerTest {
                 .content(objectMapper.writeValueAsString(intakeRequest))
         )
             .andExpect(status().isOk)
-            .andExpect(jsonPath("$.plannedAmount").value(20.0))
+            .andExpect(jsonPath("$.plannedAmount").value(20.0.toBigDecimal()))
     }
 
     @Test
@@ -210,7 +210,7 @@ class UsingsControllerTest {
         whenever(usingService.recordIntake(eq(userId), eq(drugId), any()))
             .thenThrow(ResponseStatusException(HttpStatus.BAD_REQUEST, "Exceeds planned amount"))
 
-        val intakeRequest = IntakeRequest(quantityConsumed = 100.0)
+        val intakeRequest = IntakeRequest(quantityConsumed = 100.0.toBigDecimal())
 
         mockMvc.perform(
             post("/using/drug/$drugId/intake")
